src/requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
contourpy==1.3.3
cryptography==45.0.6
cudensitymat-cu12==0.2.0
cuquantum-cu12==25.6.0
custatevec-cu12==1.9.0
cutensor-cu12==2.3.0
cutensornet-cu12==2.8.0
cycler==0.12.1
dill==0.4.0
fonttools==4.59.1
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.68.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
mpmath==1.3.0
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-nvjitlink-cu12==12.9.86
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pbr==7.0.0
pillow==11.3.0
psutil==7.0.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
pytz==2025.2
qiskit==2.1.2
qiskit-aer==0.17.1
qiskit-aer-gpu==0.15.1
qiskit-ibm-runtime==0.41.1
requests==2.32.5
requests_ntlm==1.3.0
rustworkx==0.17.1
scipy==1.16.1
seaborn==0.13.2
setuptools==78.1.1
six==1.17.0
stevedore==5.4.1
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
